+++ apps/bridge-dapp/src/hooks/index.ts
@@ -13,8 +13,10 @@ export * from './useMaxFeeInfo';
 export { default as useNavigateWithPersistParams } from './useNavigateWithPersistParams';
 export * from './useRelayerManager';
 export { default as useRelayerWithRoute } from './useRelayerWithRoute';
+export { default as useTransferFeeCalculation } from './useTransferFeeCalculation';
 export * from './useShieldedAssets';
 export * from './useSpendNotes';
 export { default as useStateWithRoute } from './useStateWithRoute';
 export * from './useTryAnotherWalletWithView';
 export { default as useTxTabFromRoute } from './useTxTabFromRoute';
+export { default as useWithdrawFeeCalculation } from './useWithdrawFeeCalculation';
