+++ apps/bridge-dapp/src/hooks/useTransferFeeCalculation.ts
@@ -1,7 +1,7 @@
 import { OptionalActiveRelayer } from '@webb-tools/abstract-api-provider/relayer/types';
 import { useWebContext } from '@webb-tools/api-provider-environment/webb-context';
 import chainsPopulated from '@webb-tools/dapp-config/chains/chainsPopulated';
-import numberToString from '@webb-tools/webb-ui-components/utils/numberToString';
+import { numberToString } from '@webb-tools/webb-ui-components';
 import { useEffect, useMemo } from 'react';
 import { BooleanParam, StringParam, useQueryParams } from 'use-query-params';
 import { formatEther, parseEther } from 'viem';
@@ -9,41 +9,41 @@ import {
   AMOUNT_KEY,
   HAS_REFUND_KEY,
   RECIPIENT_KEY,
-} from '../../../../../constants';
-import useCurrenciesFromRoute from '../../../../../hooks/useCurrenciesFromRoute';
-import {
-  MaxFeeInfoOption,
-  useMaxFeeInfo,
-} from '../../../../../hooks/useMaxFeeInfo';
+  REFUND_RECIPIENT_KEY,
+} from '../constants';
+import useCurrenciesFromRoute from './useCurrenciesFromRoute';
+import { useMaxFeeInfo, type MaxFeeInfoOption } from './useMaxFeeInfo';
 
-export default function useFeeCalculation(args: {
-  activeRelayer?: OptionalActiveRelayer;
-  recipientErrorMsg?: string;
+export default function useTransferFeeCalculation(args: {
   typedChainId?: number | null;
+  activeRelayer?: OptionalActiveRelayer;
+  refundRecipientError?: string;
+  recipientErrorMsg?: string;
 }) {
-  const { activeRelayer, recipientErrorMsg, typedChainId } = args;
+  const {
+    activeRelayer,
+    refundRecipientError,
+    recipientErrorMsg,
+    typedChainId,
+  } = args;
+
+  const { activeApi, apiConfig } = useWebContext();
 
   const [query] = useQueryParams({
     [AMOUNT_KEY]: StringParam,
-    [HAS_REFUND_KEY]: BooleanParam,
     [RECIPIENT_KEY]: StringParam,
+    [HAS_REFUND_KEY]: BooleanParam,
+    [REFUND_RECIPIENT_KEY]: StringParam,
   });
 
   const {
     [AMOUNT_KEY]: amount,
     [HAS_REFUND_KEY]: hasRefund,
+    [REFUND_RECIPIENT_KEY]: refundRecipient,
     [RECIPIENT_KEY]: recipient,
   } = query;
 
-  const chain = useMemo(() => {
-    if (typeof typedChainId === 'number') {
-      return chainsPopulated[typedChainId];
-    }
-  }, [typedChainId]);
-
-  const { activeApi, apiConfig } = useWebContext();
-
-  const { fungibleCfg, wrappableCfg } = useCurrenciesFromRoute();
+  const { fungibleCfg } = useCurrenciesFromRoute();
 
   const feeArgs = useMemo(
     () =>
@@ -59,19 +59,19 @@ export default function useFeeCalculation(args: {
     useMaxFeeInfo(feeArgs);
 
   const gasFeeInfo = useMemo(() => {
-    if (typeof feeInfo !== 'bigint') {
-      return;
+    if (typeof feeInfo === 'bigint') {
+      return feeInfo;
     }
 
-    return feeInfo;
+    return undefined;
   }, [feeInfo]);
 
   const relayerFeeInfo = useMemo(() => {
-    if (typeof feeInfo !== 'object' || feeInfo == null) {
-      return;
+    if (typeof feeInfo === 'object' && feeInfo != null) {
+      return feeInfo;
     }
 
-    return feeInfo;
+    return undefined;
   }, [feeInfo]);
 
   const refundAmount = useMemo(() => {
@@ -110,8 +110,12 @@ export default function useFeeCalculation(args: {
       return fungibleCfg?.symbol;
     }
 
-    return chain?.nativeCurrency.symbol;
-  }, [activeRelayer, chain?.nativeCurrency.symbol, fungibleCfg?.symbol]);
+    if (typeof typedChainId !== 'number') {
+      return;
+    }
+
+    return chainsPopulated[typedChainId].nativeCurrency.symbol;
+  }, [activeRelayer, fungibleCfg?.symbol, typedChainId]);
 
   const anchorId = useMemo(() => {
     if (typeof typedChainId !== 'number' || !fungibleCfg) {
@@ -121,21 +125,20 @@ export default function useFeeCalculation(args: {
     return apiConfig.getAnchorIdentifier(fungibleCfg.id, typedChainId);
   }, [apiConfig, fungibleCfg, typedChainId]);
 
-  const allInputFilled = useMemo(() => {
-    return [amount, wrappableCfg, chain, recipient, !recipientErrorMsg].every(
-      (item) => Boolean(item)
-    );
-  }, [amount, chain, recipient, recipientErrorMsg, wrappableCfg]);
-
   // Side effect for auto fetching fee info
   // when all inputs are filled and valid
   useEffect(
     () => {
-      if (!allInputFilled) {
+      if (!amount || !anchorId || typeof typedChainId !== 'number') {
         return;
       }
 
-      if (typeof typedChainId !== 'number' || !anchorId) {
+      if (!recipient || recipientErrorMsg) {
+        return;
+      }
+
+      // If refund is enabled, refund recipient must be filled and valid
+      if (hasRefund && (!refundRecipient || refundRecipientError)) {
         return;
       }
 
@@ -151,14 +154,14 @@ export default function useFeeCalculation(args: {
       fetchFeeInfo(hasRefund && hasSupport ? activeRelayer : undefined);
     },
     // prettier-ignore
-    [activeApi?.relayerManager, activeRelayer, allInputFilled, anchorId, fetchFeeInfo, hasRefund, typedChainId]
+    [activeApi?.relayerManager, activeRelayer, amount, anchorId, fetchFeeInfo, hasRefund, recipient, recipientErrorMsg, refundRecipient, refundRecipientError, typedChainId]
   );
 
   return {
     gasFeeInfo,
     isLoading,
-    relayerFeeInfo,
     refundAmount,
+    relayerFeeInfo,
     resetMaxFeeInfo,
     totalFeeToken,
     totalFeeWei,
