+++ libs/webb-ui-components/src/components/FeeDetails/RelayerFeeDetails.tsx
@@ -8,11 +8,11 @@ import { FeeDetails } from '@webb-tools/webb-ui-components/components/FeeDetails
 import type { FeeItem } from '@webb-tools/webb-ui-components/components/FeeDetails/types';
 import { Typography } from '@webb-tools/webb-ui-components/typography/Typography';
 import { formatEther } from 'viem';
-import {getRelayerFeePercentage} from './utils';
-import { useMemo } from 'react';
+import { getRelayerFeePercentage } from './utils';
+import { type FC, useMemo } from 'react';
 import { calculateTypedChainId } from '@webb-tools/sdk-core/typed-chain-id';
 
-const RelayerFeeDetails = (props: {
+interface RelayerFeeDetailsProps {
   totalFeeWei: bigint | undefined;
   totalFeeToken: string | undefined;
   gasFeeInfo: bigint | undefined;
@@ -21,18 +21,18 @@ const RelayerFeeDetails = (props: {
   srcChainCfg: ChainConfig | undefined;
   fungibleCfg: CurrencyConfig | undefined;
   activeRelayer: OptionalActiveRelayer;
-}) => {
-  const {
-    activeRelayer,
-    fungibleCfg,
-    gasFeeInfo,
-    isFeeLoading,
-    relayerFeeInfo,
-    srcChainCfg,
-    totalFeeToken,
-    totalFeeWei,
-  } = props;
+}
 
+const RelayerFeeDetails: FC<RelayerFeeDetailsProps> = ({
+  activeRelayer,
+  fungibleCfg,
+  gasFeeInfo,
+  isFeeLoading,
+  relayerFeeInfo,
+  srcChainCfg,
+  totalFeeToken,
+  totalFeeWei,
+}) => {
   const relayerFeePercentage = useMemo(() => {
     if (!activeRelayer || !srcChainCfg) {
       return;
